% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_flexalloc.R
\docType{package}
\name{flexalloc}
\alias{flexalloc}
\alias{flexalloc-package}
\title{flexalloc: Flexible allocation of work based on an auction in a crowdsourcing system}
\description{
flexalloc: Flexible allocation of work based on an auction in a crowdsourcing system
}
