% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/X1.R
\name{X1}
\alias{X1}
\title{This solves the allocation using quadratic programming and is slow.
Use X2 instead, except for checking.}
\usage{
X1(v, sumx = 1, maxx = sumx, k = 2, round = FALSE)
}
\description{
This solves the allocation using quadratic programming and is slow.
Use X2 instead, except for checking.
}
